% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labhist.R
\name{labhist}
\alias{labhist}
\title{Provide image Lab color space histgram}
\usage{
labhist(input, mode = "file", output, histgram = TRUE, resize = FALSE)
}
\arguments{
\item{input}{file}

\item{mode}{file}

\item{output}{file}

\item{histgram}{file}

\item{resize}{file}
}
\value{
image histgram of Luminance and RGB
}
\description{
Provide image Lab color space histgram
}
\examples{
labhist(input="test", mode="file", output="test")
}

