% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lrgbhist.R
\name{lrgbhist}
\alias{lrgbhist}
\title{Provide image Luminance & RGB histgram}
\usage{
lrgbhist(input, mode = "file", hist = "histgram", resize = FALSE)
}
\arguments{
\item{input}{Set file, folder or url for analysis corresponding to mode parameter.}

\item{mode}{Select a mode in all four modes. Modes are "file"(default), "url", "folder", and "scraping".}

\item{hist}{Whether histgram draw or not. Dafult is draw. However, you should set FALSE when you want fast computation.
Also this parameter is used for output name when you use folder or scraping mode (Default output name is "histgram").}

\item{resize}{file}
}
\value{
image histgram and thier descriptive stastics (Luminance & RGB)
}
\description{
Analysis an image from .jpg or .png file. Draw luminance, read, blue, and green histgram of pixel.
Calculate mean, sd, skewness, and kurtsis for each histgram.
}
\examples{
# Simple use is only set input image file.
lrgbhist(system.file("img", "newlogo.png", package="imhistR"))
# lrgbhist("img.png")  # you can use like this.

# Url mode needs to input image URL.
# Only URL tail ".jpg" or ".png" can analysis.
url <- "http://www.r-project.org/Rlogo.png"
lrgbhist(url, mode="url")
}

