% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hsbhist.R
\name{hsbhist}
\alias{hsbhist}
\title{Provide image HSB(HSV) color space histgram}
\usage{
hsbhist(input, mode = "file", output, histgram = TRUE, resize = FALSE)
}
\arguments{
\item{input}{file}

\item{mode}{file}

\item{output}{file}

\item{histgram}{file}

\item{resize}{file}
}
\value{
image histgram of Luminance and RGB
}
\description{
Provide image HSB(HSV) color space histgram
}
\examples{
hsbhist(input="test", mode="file", output="test")
}

