% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labhist.R
\name{labhist}
\alias{labhist}
\title{Provide image Lab color space histgram}
\usage{
labhist(input, mode = "file", output = input, hist = TRUE,
  resize = FALSE, endoff = FALSE, textsize = 16)
}
\arguments{
\item{input}{Set file, folder or url for image analysis corresponding to mode parameter.}

\item{mode}{Select a mode in all four modes. Modes are "file"(default), "url", "folder", and "scraping".}

\item{output}{Set the name of output histgram and data. Each mode have default output name.}

\item{hist}{Whether histgram draw or not. Dafult is draw.
However, you should set FALSE when you want fast computation for images of digital camera and smartphone.
Rendering of ggplot2 is so long for these large pixels image.}

\item{resize}{This argument is important to process many image histgram fastly.
If you set resize=1/4 or 1/8, the speed of drawing histgram is dramatically up although output values are approximation.
Resize value indicate image compression ratio. Resize is recommended when you use folder mode and want to get many histgram.}

\item{endoff}{If you want to get rid of image borders extreme value (white or black frame), you set this parameter TRUE.}

\item{textsize}{Font size of histgram caption. Dafult setting is textsize=16.}
}
\value{
image histgram and thier descriptive stastics (Lab color space). Folder and scraping mode provide a pdf file.
L range is 0-100. A and b range are -110-110. Transform Lab from RGB are based on ITU-R BT 709, using the D65 white point reference.
}
\description{
Analysis an image from  .bmp, .jpg(.jpeg) and .png file. Draw L(luminance), a(green-magenta) and b(blue-yellow) color space histgram of pixel in ggplot2. Lab color space computed from RGB value.
Calculate mean, sd, skewness, and kurtsis for each histgram.
}
\examples{
# Simple use is only set input an image file name. Japanese file name is accepted.
originaidir <- getwd()  # get current dir
setwd(system.file("extdata", package="imhistR"))  # set analysis dir
labhist("Newlogo.png", endoff=TRUE)
setwd(originaidir)      # set original dir

# labhist("yourfile.jpg")  # you can use like this after set image dir.


# Url mode needs to input image URL.
# Only URL tail ".bmp", ".jpg" or ".png" can analyze.
url <- "http://www.ess.ic.kanagawa-it.ac.jp/std_img/colorimage/Mandrill.jpg"
labhist(input=url, mode="url", output="Mandrill")


# If you have an image folder in your PC, easily analyze all images by using folder mode.
# Althogh the type of these images are limited ".bmp", ".jpg" or ".png", three type files in a folder can analyze by one command.
# Histgram is provided by a pdf file.
labhist(input=setwd(system.file("extdata", package="imhistR")), mode="folder", output="Rlogo", endoff=TRUE)

# When you analyze your picture folder, many time may have need.
# Resize give fast drawing of histgram. Compressed method is kernel.
labhist(input="folder name of iphone picture", mode="folder", resize=1/4)


# Web scraping from google image search is conducted by scraping mode. Twenty images were automatically downloaded and analyzed.
# So many scraping should avoid in order to conform web manner.
# If you already scraping image by other function of this package, you should use folder mode.
# Histgram is provided by a pdf file.
url <- "url from google image search of xxx"  # This package does not provide the way to scraping other web pages
labhist(input=url, mode="scraping")
}

