% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labhist.R
\name{labhist}
\alias{labhist}
\title{Provide image Lab color space histgram}
\usage{
labhist(input, mode = "file", hist = "histgram", resize = FALSE)
}
\arguments{
\item{input}{Set file, folder or url for image analysis corresponding to mode parameter.}

\item{mode}{Select a mode in all four modes. Modes are "file"(default), "url", "folder", and "scraping".}

\item{hist}{Whether histgram draw or not. Dafult is draw. However, you should set FALSE when you want fast computation.
Also this parameter is used for output name when you use folder or scraping mode (Default output name is "histgram").}

\item{resize}{file}
}
\value{
image histgram and thier descriptive stastics (Lab color space). L range is 0-100. A and b range are -110-110.
Transform Lab from RGB are based on ITU-R BT 709, using the D65 white point reference.
}
\description{
Analysis an image from .jpg or .png file. Draw L(luminance), a(green-magenta) and b(blue-yellow) color space histgram of pixel in ggplot2. Lab color space computed from RGB value.
Calculate mean, sd, skewness, and kurtsis for each histgram.
}
\examples{
# Simple use is only set input image file.
labhist(system.file("img", "newlogo.png", package="imhistR"))
# labhist("img.png")  # you can use like this.

# Url mode needs to input image URL.
# Only URL tail ".jpg" or ".png" can analysis.
url <- "http://www.r-project.org/Rlogo.png"
labhist(url, mode="url")
}

