% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lrgbhist.R
\name{lrgbhist}
\alias{lrgbhist}
\title{Provide image Luminance & RGB histgram}
\usage{
lrgbhist(input, mode = "file", hist = "histgram", resize = FALSE)
}
\arguments{
\item{input}{file}

\item{mode}{file}

\item{hist}{file}

\item{resize}{file}
}
\value{
image histgram of Luminance and RGB
}
\description{
Most data operations are useful done on groups defined by variables in the
the dataset. The \code{group_by} function takes an existing tbl
and converts it into a grouped tbl where operations are performed
"by group".
}
\examples{
lrgbhist("test")
}

